+++ backend/canisters/community/api/src/queries/selected_updates_v2.rs
@@ -1,6 +1,6 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::{AccessRules, CommunityMember, TimestampMillis, UserGroupDetails, UserId, VersionedRules};
+use types::{CommunityMember, Rules, TimestampMillis, UserGroupDetails, UserId, VersionedRules};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -24,9 +24,9 @@ pub struct SuccessResult {
     pub blocked_users_added: Vec<UserId>,
     pub blocked_users_removed: Vec<UserId>,
     pub invited_users: Option<Vec<UserId>>,
-    // TODO: remove this field once the website is using `access_rules` instead
-    pub rules: Option<AccessRules>,
-    pub access_rules: Option<VersionedRules>,
+    // TODO: remove this field once the website is using `chat_rules` instead
+    pub rules: Option<Rules>,
+    pub chat_rules: Option<VersionedRules>,
     pub user_groups: Vec<UserGroupDetails>,
     pub user_groups_deleted: Vec<u32>,
 }