+++ backend/canisters/community/api/src/updates/create_channel.rs
@@ -1,8 +1,8 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use types::{
-    AccessGate, AccessRules, ChannelId, Document, FieldTooLongResult, FieldTooShortResult, GroupPermissions, GroupSubtype,
-    Milliseconds,
+    AccessGate, ChannelId, Document, FieldTooLongResult, FieldTooShortResult, GroupPermissions, GroupSubtype, Milliseconds,
+    Rules,
 };
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -10,7 +10,7 @@ pub struct Args {
     pub is_public: bool,
     pub name: String,
     pub description: String,
-    pub rules: AccessRules,
+    pub rules: Rules,
     pub subtype: Option<GroupSubtype>,
     pub avatar: Option<Document>,
     pub history_visible_to_new_joiners: bool,