+++ backend/canisters/local_group_index/api/src/updates/c2c_create_group.rs
@@ -1,6 +1,6 @@
 use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
-use types::{AccessGate, AccessRules, ChatId, Document, GroupPermissions, GroupSubtype, Milliseconds, UserId};
+use types::{AccessGate, ChatId, Document, GroupPermissions, GroupSubtype, Milliseconds, Rules, UserId};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -9,7 +9,7 @@ pub struct Args {
     pub is_public: bool,
     pub name: String,
     pub description: String,
-    pub rules: AccessRules,
+    pub rules: Rules,
     pub subtype: Option<GroupSubtype>,
     pub avatar: Option<Document>,
     pub history_visible_to_new_joiners: bool,